+++ test_requirements.txt
@@ -24,3 +24,4 @@ six==1.11.0
 tox
 
 git+https://github.com/edx/xblock-utils.git@v1.0.5#egg=xblock-utils==1.0.5
+git+https://github.com/edx/xblock-sdk.git#egg=xblock-sdk
